refactor: Update repeated-panel layout in PageWidget
This commit updates the layout of the repeated-panel in the PageWidget. It introduces a new feature that allows users to dynamically add repeated panels by clicking on the add button. The repeated panels are now displayed in a row with a grey background color. Each added panel contains a set of fields wrapped in an AppCard. This enhancement improves the user experience and provides a more intuitive way to interact with the form.